Added condition to delete from store only when exists. Added case of saving the bundle to the store when uploading from url. Added checks in plugin tests

+++ app/plugin_install.go
@@ -140,8 +140,10 @@ func (a *App) removePlugin(id string) *model.AppError {
 
 	// Remove bundle from the file store.
 	storePluginFileName := filepath.Join("./plugins", manifest.Id) + ".tar.gz"
-	if err := a.RemoveFile(storePluginFileName); err != nil {
-		return model.NewAppError("removePlugin", "app.plugin.remove_bundle.app_error", nil, err.Error(), http.StatusInternalServerError)
+	if  bundleExist, _ := a.FileExists(storePluginFileName); bundleExist {
+		if err := a.RemoveFile(storePluginFileName); err != nil {
+			return model.NewAppError("removePlugin", "app.plugin.remove_bundle.app_error", nil, err.Error(), http.StatusInternalServerError)
+		}
 	}
 
 	return nil
